pyproject.toml
[project]
name = "sumstats-ratio-qtl"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pandas-stubs==2.3.2.250926",
    "scipy-stubs==1.16.3.0",
]
